Refactor: styleConstant 파일에 커스텀 Full class name을 보관하도록 개선
